Refactor code into a new custom view "VrView".  Move all gles helper code to sub package "vr".  Set two surfaces to correct aspect for video format.
