+++ app/src/main/java/com/fpvout/digiview/vr/gles/EglSurfaceBase.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.fpvout.digiview.gles;
+package com.fpvout.digiview.vr.gles;
 
 import android.graphics.Bitmap;
 import android.opengl.EGL14;
