+++ app/src/main/java/com/fpvout/digiview/vr/gles/GlUtil.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.fpvout.digiview.gles;
+package com.fpvout.digiview.vr.gles;
 
 import android.opengl.GLES20;
 import android.opengl.GLES30;
